Badly not working progress on FSM refactor with Mel::SceneFSM
