+++ mobilis.gemspec
@@ -34,7 +34,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency "awesome_print"
   spec.add_dependency "optimist"
   spec.add_dependency "pry"
-  spec.add_dependency "state_machine"
+  spec.add_dependency "state_machines"
   spec.add_dependency "table_print"
   spec.add_dependency "tty-prompt"
